AI/diffusionGan/requirements.txt
torch==2.1.0
torchvision==0.15.0
Pillow==9.5.0
matplotlib==3.8.0
numpy==1.24.0

opencv-python
opencv-python-headless
